% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JavaScript}
\alias{JavaScript}
\alias{jseval}
\alias{jsvalidate}
\title{Validate and evaluate JavaScript code}
\usage{
jseval(code)

jsvalidate(code)
}
\arguments{
\item{code}{A string with JavaScript code}
}
\value{
Console output.
}
\description{
The \code{jsvalidate} function tests if a string is valid JavaScript.
The \code{jseval} function evaluates a JavaScript string and returns
console output, equivalent to \code{eval()} in JavaScript. These
functions are stateless: their context is destroyed after the evaluation
has completed.
}
\examples{
# Evaluate JavaScript code
jseval("JSON.stringify({x:Math.random()})")
jseval("(function(x){return x+1;})(123)")
jseval(c("foo = 123", "bar = 456", "foo + bar"))

# Load a library (doesn't do anything yet)
underscore <- system.file("js/underscore.js", package="V8")
jseval(readLines(underscore))

# Cannot define anonymous function in global scope
jsvalidate("function(x){2*x}") #FALSE
jsvalidate("function foo(x){2*x}") #TRUE
jsvalidate("foo = function(x){2*x}") #TRUE
}

