% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JS.R
\name{JS}
\alias{JS}
\title{Mark character strings as literal JavaScript code}
\usage{
JS(...)
}
\arguments{
\item{...}{character vectors as the JavaScript source code (all arguments
will be pasted into one character string)}
}
\description{
This function \code{JS()} marks character vectors with a special class, so
that it will be treated as literal JavaScript code. It was copied from the
htmlwidgets package, and does exactly the same thing.
}
\examples{
ct <- new_context()
ct$eval("1+1")
ct$eval(JS("1+1"))
ct$assign("test", JS("2+3"))
ct$get("test")
}
\author{
Yihui Xie
}

