% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jseval}
\alias{jseval}
\title{Evaluate JavaScript string}
\usage{
jseval(code)
}
\arguments{
\item{code}{a string with JavaScript code.}
}
\description{
This function works like calling \code{eval()} in JavaScript.
}
\examples{
jseval("JSON.stringify({x:Math.random()});")
jseval("(function(x){return x+1;})(123);")
}

