% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasm.R
\name{wasm}
\alias{wasm}
\title{Run WebAssembly}
\usage{
wasm(data)
}
\arguments{
\item{data}{either raw vector or file path with the binary wasm program}
}
\description{
Load a WebAssembly program. Returns a list of exported functions.
}
\examples{
# Load example wasm program
if(engine_info()$version > 6){
instance <- wasm(system.file('wasm/add.wasm', package = 'V8'))
instance$exports$add(12, 30)
}
}
