% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JavaScript}
\alias{JavaScript}
\alias{jseval}
\alias{jsvalidate}
\title{Validate and evaluate JavaScript code}
\usage{
jseval(code)

jsvalidate(code)
}
\arguments{
\item{code}{A string with JavaScript code}
}
\value{
Console output.
}
\description{
The \code{jsvalidate} function tests if a string is valid JavaScript.
The \code{jseval} function evaluates a JavaScript string and returns
console output, equivalent to \code{eval()} in JavaScript. These
functions are stateless: their context is destroyed after the evaluation
has completed.
}
\examples{
jseval("JSON.stringify({x:Math.random()})")
jseval("(function(x){return x+1;})(123)")
jseval("foo = 123; bar = 456; foo + bar")

# Cannot define anonymous function in global namespace
jsvalidate("function(x){2*x}")
jsvalidate("function foo(x){2*x}")
jsvalidate("foo = function(x){2*x}")
}

