% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new_context}
\alias{V8}
\alias{new_context}
\title{Embedded JavaScript}
\usage{
new_context()
}
\description{
The \code{jsvalidate} function tests if a string is valid JavaScript.
The \code{jseval} function evaluates a JavaScript string and returns
console output, equivalent to \code{eval()} in JavaScript. These
functions are stateless: their context is destroyed after the evaluation
has completed.
}
\examples{
# Create a new context
ct <- new_context();

# Evaluate some code
ct$eval("foo=123")
ct$eval("bar=456")
ct$eval("foo+bar")

# Functions and closures
ct$eval("JSON.stringify({x:Math.random()})")
ct$eval("(function(x){return x+1;})(123)")

# Objects (via JSON only)
ct$assign("mydata", mtcars)
ct$get("mydata")

# Validate syntax without evaluating
ct$validate("function foo(x){2*x}") #TRUE
ct$validate("foo = function(x){2*x}") #TRUE
ct$validate("function(x){2*x}") #FALSE

# Load a JavaScript library
ct$source(system.file("js/underscore.js", package="V8"))

# Call a function
}

