% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new_context}
\alias{V8}
\alias{new_context}
\title{Run JavaScript in a V8 context}
\usage{
new_context()
}
\description{
A \emph{context} is an execution environment that allows separate, unrelated,
JavaScript code to run in a single instance of V8. You must explicitly specify
the context in which you want any JavaScript code to be run.
}
\examples{
# Create a new context
ct <- new_context();

# Evaluate some code
ct$eval("foo=123")
ct$eval("bar=456")
ct$eval("foo+bar")

# Functions and closures
ct$eval("JSON.stringify({x:Math.random()})")
ct$eval("(function(x){return x+1;})(123)")

# Objects (via JSON only)
ct$assign("mydata", mtcars)
ct$get("mydata")

# Validate syntax without evaluating
ct$validate("function foo(x){2*x}") #TRUE
ct$validate("foo = function(x){2*x}") #TRUE
ct$validate("function(x){2*x}") #FALSE

# Load a JavaScript library
ct$source(system.file("js/underscore.js", package="V8"))

# Call a function
ct$call("function(x){return x*2}", 123)
}

